update test to check first and second byte
