+++ Readme.md
@@ -8,7 +8,7 @@ Please do not attempt to use this in production. It's simply not ready.
 # Min Requirements
 - Install Google Chrome otherwise web tests will fail. If you do not want to install Google Chrome, add the environment variable `IGNORE_BROWSER_TESTS=true`
 
-# Test Status (>1528 tests as of now)
+# Test Status (>2793 tests as of now)
 Currently most of the tests are unit tests and there are some integration tests validating that we can connect to a mosquitto server
 Check the latest build gradle scan at the end of the travis build to see how many tests we are currently running.
